Specialized macros for autoRefine{V,T}
This specializes the macro implementations for `autoRefineV` and
`autoRefineT`. Both methods were defined in terms of the generic
`RefineMacro.impl` which works for any `RefType[F]`. This generic
implementation has the drawback that it needs to `eval` the implicit
`RefType` parameter at compile-time. Since `autoRefine{V,T}` already
know which `RefType` instance they will use, it is unnecessary to `eval`
them in the macro.

Here are the benchmark results for master and this change:
```
master:

[info] Result "eu.timepit.refined.benchmark.RefineMacroBenchmark.autoRefineV_PosInt":
[info]   0.549 ±(99.9%) 0.004 s/op [Average]
[info]   (min, avg, max) = (0.520, 0.549, 0.593), stdev = 0.015
[info]   CI (99.9%): [0.546, 0.553] (assumes normal distribution)
[info] # Run complete. Total time: 00:17:45
[info] Benchmark                                Mode  Cnt  Score   Error Units
[info] RefineMacroBenchmark.autoRefineV_PosInt  avgt  200  0.549 ± 0.004 s/op

impl_Refined:

[info] Result "eu.timepit.refined.benchmark.RefineMacroBenchmark.autoRefineV_PosInt":
[info]   0.403 ±(99.9%) 0.004 s/op [Average]
[info]   (min, avg, max) = (0.377, 0.403, 0.471), stdev = 0.017
[info]   CI (99.9%): [0.399, 0.407] (assumes normal distribution)
[info] # Run complete. Total time: 00:09:07
[info] Benchmark                                Mode  Cnt  Score   Error Units
[info] RefineMacroBenchmark.autoRefineV_PosInt  avgt  200  0.403 ± 0.004 s/op
```

That is a decrease in compilation time of ~26%. This is first PR towards
reduced compilation times: https://github.com/fthomas/refined/issues/328