Test that instances don't need to imported explicitly
This tests that `Validate` and `Inference` instances are in the implicit
scope of their predicates and do not need to be imported explicitly.

I conducted https://github.com/fthomas/refined/pull/392 under the
premise that `Validate` and `Inference` instances were not in the implicit
scope of their predicates and using them would always require an
explicit import. For example, to refine with the `numeric.Positive`
predicate I thought a `numeric._` wildcard import was always necessary.
This tests show that this is not the case.