Make Adjacent a subclass of Ordering
This PR

1. makes `Adjacent` a subclass of `Ordering`
2. adds `next{Up,Down}OrNone` to `Adjacent`
3. changes `{Min,Max}.findValid` to throw an exception instead of
   looping forever if it fails to find a next value satisfying the
   predicate

Making `Adjacent` a subclass of `Ordering` was not required technically
but is the right thing to do conceptually. The laws of `Adjacent` are
`nextUp(x) >= x` and `nextDown(x) <= x` for all `x`. To formulate these
laws we need `>=` and `<=` and hence any `Adjacent` needs to be
`Ordering`.