Add a TrimmedString.trim method
This is somewhat similar to `FiniteStringOps.truncate` in that it's
guaranteed to turn any `String` into a `TrimmedString`.

There are some funny characters that get removed by `String.trim` but
don't match on `\s` in regular expressions. I've added an example to
demonstrate that these characters can lead to a `String` being "trimmed"
a bit more than would technically be necessary to match the
`TrimmedString` regex.